address review(kyle) feedback
1. merge QuickViewTabHelper into existed BraveSearechTabHelper
2. passing any Profile to QVController instead of some Tab and a PrivateBrowsingManager
3. some other nits
